+++ src/abi.ts
@@ -1,9 +1,8 @@
-import { int2Asm, bsv, arrayTypeAndSize, genLaunchConfigFile, getStructNameByType, isArrayType, isStructType, checkArray, flatternArray, typeOfArg, flatternStruct, createStruct, createArray, asm2ScryptType } from './utils';
+import { int2Asm, bsv, genLaunchConfigFile, getStructNameByType, isArrayType, isStructType, checkArray, flatternArray, typeOfArg, deserializeArgfromASM, createStruct, createArray, asm2ScryptType, bin2num } from './utils';
 import { AbstractContract, TxContext, VerifyResult, AsmVarValues } from './contract';
-import { ScryptType, Bool, Int, SingletonParamType, SupportedParamType, Struct, TypeResolver, PubKey, Sig, Ripemd160, Sha1, Sha256, SigHashType, SigHashPreimage, OpCodeType } from './scryptTypes';
+import { ScryptType, Bool, Int, SupportedParamType, Struct, TypeResolver, VariableType } from './scryptTypes';
 import { ABIEntityType, ABIEntity, ParamEntity } from './compilerWrapper';
-import { buildContractCodeASM, readState } from './internal';
-import { bin2num, Bytes, VariableType } from '.';
+import { buildContractCodeASM, flatternArgs, flatternStateArgs, readState } from './internal';
 
 export interface Script {
   toASM(): string;
@@ -76,7 +75,7 @@ export class FunctionCall {
       contract: AbstractContract;
       lockingScriptASM?: string;
       unlockingScriptASM?: string;
-      params: SupportedParamType[];
+      args: Arguments;
     }
   ) {
 
@@ -86,22 +85,7 @@ export class FunctionCall {
 
     this.contract = binding.contract;
 
-
-    this.args = Object.getPrototypeOf(this.contract).constructor.abi.filter((entity: ABIEntity) => {
-      if ('constructor' === methodName) {
-        return entity.type === 'constructor';
-      }
-      return entity.name === methodName;
-    }).map((entity: ABIEntity) => {
-      return entity.params.map((param, index) => {
-        return {
-          name: param.name,
-          type: param.type,
-          state: param.state || false,
-          value: binding.params[index]
-        };
-      });
-    }).flat(1);
+    this.args = binding.args;
 
     if (binding.lockingScriptASM) {
       this._lockingScriptAsm = binding.lockingScriptASM;
@@ -177,93 +161,91 @@ export class ABICoder {
 
   constructor(public abi: ABIEntity[], public finalTypeResolver: TypeResolver) { }
 
+  checkArgs(fname: string, params: ParamEntity[], ...args: SupportedParamType[]): void {
+
+    if (args.length !== params.length) {
+      throw new Error(`wrong number of arguments for #${fname}, expected ${params.length} but got ${args.length}`);
+    }
+
+    params.map(p => ({
+      name: p.name,
+      type: this.finalTypeResolver(p.type)
+    })).forEach((param, index) => {
+      const arg = args[index];
+      if (isArrayType(param.type)) {
+        if (!checkArray(arg as SupportedParamType[], param.type)) {
+          throw new Error(`The type of parameter ${param.name} is wrong, should be ${param.type}`);
+        }
+      } else {
+        const scryptType = typeOfArg(arg);
+        if (scryptType != param.type) {
+          const expected = isStructType(param.type) ? getStructNameByType(param.type) : param.type;
+          const got = isStructType(scryptType) ? getStructNameByType(scryptType) : scryptType;
+          throw new Error(`The type of parameter ${param.name} is wrong, expected ${expected} but got ${got}`);
+        }
+      }
+    });
+  }
 
   encodeConstructorCall(contract: AbstractContract, asmTemplate: string, ...args: SupportedParamType[]): FunctionCall {
 
     const constructorABI = this.abi.filter(entity => entity.type === ABIEntityType.CONSTRUCTOR)[0];
     const cParams = constructorABI?.params || [];
 
-    if (args.length !== cParams.length) {
-      throw new Error(`wrong number of arguments for #constructor, expected ${cParams.length} but got ${args.length}`);
-    }
+    this.checkArgs('constructor', cParams, ...args);
 
     // handle array type
-    const cParams_: Array<ParamEntity> = [];
-    const args_: SupportedParamType[] = [];
-    cParams.map(p => ({
-      name: p.name,
-      type: this.finalTypeResolver(p.type),
-      state: p.state
-    })).forEach((param, index) => {
+    const flatteredArgs = flatternArgs(cParams.map((p, index) => (Object.assign({ ...p }, {
+      state: p.state,
+      value: args[index]
+    }))), this.finalTypeResolver);
 
-      if (isArrayType(param.type)) {
-        const arg = args[index] as SupportedParamType[];
-        if (checkArray(arg, param.type)) {
-          // flattern array
-          flatternArray(arg, param.name, param.type).forEach((e) => {
-            cParams_.push({ name: e.name, type: this.finalTypeResolver(e.type), state: param.state });
-            args_.push(e.value);
-          });
-        } else {
-          throw new Error(`constructor ${index}-th parameter should be ${param.type}`);
-        }
 
-      } else if (isStructType(param.type)) {
-        const arg = args[index] as Struct;
 
-        if (param.type != arg.finalType) {
-          throw new Error(`expect struct ${getStructNameByType(param.type)} but got struct ${arg.type}`);
-        }
-
-        flatternStruct(arg, param.name).forEach(v => {
-          cParams_.push({ name: `${v.name}`, type: this.finalTypeResolver(v.type), state: param.state });
-          args_.push(v.value);
-        });
+    flatteredArgs.forEach(arg => {
+      if (!asmTemplate.includes(`$${arg.name}`)) {
+        throw new Error(`abi constructor params mismatch with args provided: missing ${arg.name} in ASM tempalte`);
       }
-      else {
-        const arg = args[index] as SupportedParamType;
-        cParams_.push(param);
-        args_.push(arg);
-      }
-    });
-
-
-
-    cParams_.forEach((param, index) => {
-      if (!asmTemplate.includes(`$${param.name}`)) {
-        throw new Error(`abi constructor params mismatch with args provided: missing ${param.name} in ASM tempalte`);
-      }
-      if (param.state) {
+      if (arg.state) {
         //if param is state , we use default value to new contract.
-        if (param.type === VariableType.INT || param.type === VariableType.PRIVKEY) {
-          contract.asmTemplateArgs.set(`$${param.name}`, 'OP_0');
-        } else if (param.type === VariableType.BOOL) {
-          contract.asmTemplateArgs.set(`$${param.name}`, 'OP_TRUE');
-        } else if (param.type === VariableType.BYTES
-          || param.type === VariableType.PUBKEY
-          || param.type === VariableType.SIG
-          || param.type === VariableType.RIPEMD160
-          || param.type === VariableType.SHA1
-          || param.type === VariableType.SHA256
-          || param.type === VariableType.SIGHASHTYPE
-          || param.type === VariableType.SIGHASHPREIMAGE
-          || param.type === VariableType.OPCODETYPE) {
-          contract.asmTemplateArgs.set(`$${param.name}`, '00');
+        if (arg.type === VariableType.INT || arg.type === VariableType.PRIVKEY) {
+          contract.asmTemplateArgs.set(`$${arg.name}`, 'OP_0');
+        } else if (arg.type === VariableType.BOOL) {
+          contract.asmTemplateArgs.set(`$${arg.name}`, 'OP_TRUE');
+        } else if (arg.type === VariableType.BYTES
+          || arg.type === VariableType.PUBKEY
+          || arg.type === VariableType.SIG
+          || arg.type === VariableType.RIPEMD160
+          || arg.type === VariableType.SHA1
+          || arg.type === VariableType.SHA256
+          || arg.type === VariableType.SIGHASHTYPE
+          || arg.type === VariableType.SIGHASHPREIMAGE
+          || arg.type === VariableType.OPCODETYPE) {
+          contract.asmTemplateArgs.set(`$${arg.name}`, '00');
         } else {
-          throw new Error(`param ${param.name} has unknown type ${param.type}`);
+          throw new Error(`param ${arg.name} has unknown type ${arg.type}`);
         }
 
       } else {
-        contract.asmTemplateArgs.set(`$${param.name}`, this.encodeParam(args_[index], param));
+        contract.asmTemplateArgs.set(`$${arg.name}`, this.encodeParam(arg.value, arg));
       }
-
-
 
     });
 
     contract.asmTemplateArgs.set('$__codePart__', 'OP_0');
 
-    return new FunctionCall('constructor', { contract, lockingScriptASM: buildContractCodeASM(contract.asmTemplateArgs, asmTemplate), params: args });
+
+    return new FunctionCall('constructor', {
+      contract,
+      lockingScriptASM: buildContractCodeASM(contract.asmTemplateArgs, asmTemplate),
+      args: cParams.map((param, index) => ({
+        name: param.name,
+        type: param.type,
+        state: param.state,
+        value: args[index]
+      }))
+    });
+
   }
 
   encodeConstructorCallFromRawHex(contract: AbstractContract, asmTemplate: string, raw: string): FunctionCall {
@@ -302,41 +284,22 @@ export class ABICoder {
       name: p.name,
       value: undefined,
       state: p.state
-    })).map(param => {
+    })).map(arg => {
 
-      let value;
-
-      if (isStructType(param.type)) {
-
-        const stclass = contract.getTypeClassByType(getStructNameByType(param.type));
-
-        value = createStruct(contract, stclass as typeof Struct, param.name, contract.asmTemplateArgs, this.finalTypeResolver);
-      } else if (isArrayType(param.type)) {
-
-        value = createArray(contract, param.type, param.name, contract.asmTemplateArgs, this.finalTypeResolver);
-
-      } else {
-        value = asm2ScryptType(param.type, contract.asmTemplateArgs.get(`$${param.name}`));
-      }
-
-      return {
-        name: param.name,
-        type: this.finalTypeResolver(param.type),
-        state: param.state,
-        value: value
-      };
-
+      deserializeArgfromASM(contract, arg, contract.asmTemplateArgs);
+      return arg;
     });
 
 
     if (AbstractContract.isStateful(contract)) {
 
+      const flatteredArgs = flatternStateArgs(args, this.finalTypeResolver);
 
       const scriptHex = script.toHex();
       const metaScript = script.toHex().substr(scriptHex.length - 10, 10);
       const version = bin2num(metaScript.substr(metaScript.length - 2, 2)) as number;
       const stateLen = bin2num(metaScript.substr(0, 8)) as number;
-
+      const stateAsmTemplateArgs: Map<string, string> = new Map();
       switch (version) {
         case 0:
           {
@@ -348,48 +311,33 @@ export class ABICoder {
 
             const br = new bsv.encoding.BufferReader(stateHex);
 
-            args.forEach((arg) => {
-              if (arg.state) {
-                if (arg.type === VariableType.BOOL) {
-                  const opcodenum = br.readUInt8();
-                  arg.value = new Bool(opcodenum === 1);
+            flatteredArgs.forEach((arg) => {
+              if (arg.type === VariableType.BOOL) {
+                const opcodenum = br.readUInt8();
+                stateAsmTemplateArgs.set(`$${arg.name}`, opcodenum === 1 ? '01' : '00');
+              } else {
+                const { data } = readState(br);
+                if (arg.type === VariableType.INT || arg.type === VariableType.PRIVKEY) {
+                  stateAsmTemplateArgs.set(`$${arg.name}`, new Int(bin2num(data)).toASM());
                 } else {
-                  const { data } = readState(br);
-                  if (arg.type === VariableType.INT || arg.type === VariableType.PRIVKEY) {
-                    arg.value = new Int(bin2num(data));
-                  } else if (arg.type === VariableType.BYTES) {
-                    arg.value = new Bytes(data);
-                  } else if (arg.type === VariableType.PUBKEY) {
-                    arg.value = new PubKey(data);
-                  } else if (arg.type === VariableType.SIG) {
-                    arg.value = new Sig(data);
-                  } else if (arg.type === VariableType.RIPEMD160) {
-                    arg.value = new Ripemd160(data);
-                  } else if (arg.type === VariableType.SHA1) {
-                    arg.value = new Sha1(data);
-                  } else if (arg.type === VariableType.SHA256) {
-                    arg.value = new Sha256(data);
-                  } else if (arg.type === VariableType.SIGHASHTYPE) {
-                    arg.value = new SigHashType(bin2num(data) as number);
-                  } else if (arg.type === VariableType.SIGHASHPREIMAGE) {
-                    arg.value = new SigHashPreimage(data);
-                  } else if (arg.type === VariableType.OPCODETYPE) {
-                    arg.value = new OpCodeType(data);
-                  } else {
-                    //
-                    throw new Error('unsupport type ' + arg.type);
-                  }
+                  stateAsmTemplateArgs.set(`$${arg.name}`, data);
                 }
               }
             });
+
           }
           break;
       }
 
+      args.filter(a => a.state).forEach(arg => {
+
+        deserializeArgfromASM(contract, arg, stateAsmTemplateArgs);
+      });
+
       lsASM = buildContractCodeASM(contract.asmTemplateArgs, asmTemplate);
     }
 
-    return new FunctionCall('constructor', { contract, lockingScriptASM: lsASM, params: args.map(arg => arg.value) });
+    return new FunctionCall('constructor', { contract, lockingScriptASM: lsASM, args: args });
 
   }
 
@@ -397,9 +345,7 @@ export class ABICoder {
 
     for (const entity of this.abi) {
       if (entity.name === name) {
-        if (entity.params.length !== args.length) {
-          throw new Error(`wrong number of arguments for #${name}, expected ${entity.params.length} but got ${args.length}`);
-        }
+        this.checkArgs(name, entity.params, ...args);
         let asm = this.encodeParams(args, entity.params.map(p => ({
           name: p.name,
           type: this.finalTypeResolver(p.type),
@@ -410,7 +356,14 @@ export class ABICoder {
           const pubFuncIndex = entity.index;
           asm += ` ${int2Asm(pubFuncIndex.toString())}`;
         }
-        return new FunctionCall(name, { contract, unlockingScriptASM: asm, params: args });
+        return new FunctionCall(name, {
+          contract, unlockingScriptASM: asm, args: entity.params.map((param, index) => ({
+            name: param.name,
+            type: param.type,
+            state: false,
+            value: args[index]
+          }))
+        });
       }
     }
 
@@ -422,72 +375,35 @@ export class ABICoder {
   }
 
   encodeParamArray(args: SupportedParamType[], arrayParam: ParamEntity): string {
-    if (args.length === 0) {
-      throw new Error('Empty array not allowed');
-    }
-
-    const t = typeof args[0];
-
-    if (!args.every(arg => typeof arg === t)) {
-      throw new Error('Array arguments are not of the same type');
-    }
-
-    if (checkArray(args, arrayParam.type)) {
-      return flatternArray(args, arrayParam.name, arrayParam.type).map(arg => {
-        return this.encodeParam(arg.value, { name: arg.name, type: this.finalTypeResolver(arg.type), state: arrayParam.state });
-      }).join(' ');
-    } else {
-      throw new Error(`checkArray ${arrayParam.type} fail`);
-    }
+    return flatternArray(args, arrayParam.name, arrayParam.type).map(arg => {
+      return this.encodeParam(arg.value, { name: arg.name, type: this.finalTypeResolver(arg.type), state: arrayParam.state });
+    }).join(' ');
   }
 
 
   encodeParam(arg: SupportedParamType, paramEntity: ParamEntity): string {
 
     if (isArrayType(paramEntity.type)) {
-      if (Array.isArray(arg)) {
-        return this.encodeParamArray(arg, paramEntity);
-      } else {
-        const scryptType = typeOfArg(arg);
-        throw new Error(`expect param ${paramEntity.name} as ${paramEntity.type} but got ${scryptType}`);
-      }
+      return this.encodeParamArray(arg as SupportedParamType[], paramEntity);
     }
 
-    if (isStructType(paramEntity.type)) {
-
-      if (Struct.isStruct(arg)) {
-        const argS = arg as Struct;
-        if (paramEntity.type != argS.finalType) {
-          throw new Error(`expect struct ${getStructNameByType(paramEntity.type)} but got struct ${argS.type}`);
-        }
-      } else {
-        const scryptType = (arg as ScryptType).type;
-        throw new Error(`expect param ${paramEntity.name} as struct ${getStructNameByType(paramEntity.type)} but got ${scryptType}`);
-      }
-    }
-
-
-    const scryptType = typeOfArg(arg);
-    if (scryptType != paramEntity.type) {
-      throw new Error(`wrong argument type, expected ${paramEntity.type} but got ${scryptType}`);
+    if (arg instanceof ScryptType) {
+      return arg.toASM();
     }
 
     const typeofArg = typeof arg;
 
     if (typeofArg === 'boolean') {
       arg = new Bool(arg as boolean);
-    }
-
-    if (typeofArg === 'number') {
+    } else if (typeofArg === 'number') {
       arg = new Int(arg as number);
-    }
-
-    if (typeofArg === 'bigint') {
+    } else if (typeofArg === 'bigint') {
       arg = new Int(arg as bigint);
-    }
-
-    if (typeof arg === 'string') {
+    } else if (typeof arg === 'string') {
       arg = new Int(arg as string);
+    } else {
+      //we call checkArg before encodeParam, shouldn't get here under normal circumstances
+      throw new Error(`The value of parameter ${paramEntity.name} is unknown type: ${typeofArg}`);
     }
 
     return (arg as ScryptType).toASM();